training: integrate precision policy, schedulers, and reproducibility seeds
- Wire precision policy into trainer with FP32 params + BF16 autocast
- Use compute_scheduler_steps for warmup/total step calculation
- Add deterministic seeding for reproducibility
- Add gradient accumulation step support
- Enhance JointInferenceLoss with imputed weight handling
- Add Kalman filter tests for state estimation

+++ tests/test_training_resume.py
@@ -138,7 +138,7 @@ def test_resume_from_checkpoint_loads_state(tmp_path) -> None:
     )
     trainer = _make_trainer_stub(config)
     trainer.model_id = "run_003"
-    trainer.model = torch.nn.Linear(2, 2)
+    trainer.model = torch.nn.Linear(2, 2, dtype=torch.float32)
     trainer.optimizer = torch.optim.SGD(trainer.model.parameters(), lr=0.1)
     trainer.scheduler = None
     trainer.best_val_loss = float("inf")
